Remove old signature view modules
