+++ backend/signature/views/notification.py
@@ -1,0 +1,15 @@
+from rest_framework import viewsets
+from rest_framework.permissions import IsAuthenticated
+
+from ..models import NotificationPreference
+from ..serializers import NotificationPreferenceSerializer
+
+class NotificationPreferenceViewSet(viewsets.ModelViewSet):
+    serializer_class = NotificationPreferenceSerializer
+    permission_classes = [IsAuthenticated]
+
+    def get_queryset(self):
+        return NotificationPreference.objects.filter(user=self.request.user)
+
+    def perform_create(self, serializer):
+        serializer.save(user=self.request.user)
